IEI-175871 DCM4CHE3 does not support elements> 2GB in length

+++ dcm4che-core/src/main/java/org/dcm4che3/io/stream/CloseableImageInputStreamAdapter.java
@@ -96,7 +96,7 @@ public class CloseableImageInputStreamAdapter extends InputStream {
 
     @Override
     public long skip(long n) throws IOException {
-        return iis.skipBytes((int) n);
+        return iis.skipBytes(n);
     }
 
     @Override
